# Produce NoSuchTable Pact error, and clean up ChainwebPactDb