--- a/test/unit/Chainweb/Test/Pact4/Checkpointer.hs
+++ b/test/unit/Chainweb/Test/Pact4/Checkpointer.hs
@@ -66,6 +66,7 @@ import Chainweb.Version
 import Chainweb.Test.Orphans.Internal ({- Arbitrary BlockHash -})
 import Chainweb.Pact.Backend.Types
 import qualified Chainweb.Pact.PactService.Checkpointer.Internal as Checkpointer
+import qualified Chainweb.Pact5.Backend.ChainwebPactDb as Pact5
 
 -- -------------------------------------------------------------------------- --
 -- Tests
@@ -761,7 +762,7 @@ simpleBlockEnvInit
     -> (PactDb (BlockEnv logger) -> BlockEnv logger -> (MVar (BlockEnv logger) -> IO ()) -> IO a)
     -> IO a
 simpleBlockEnvInit logger f = withTempSQLiteConnection chainwebPragmas $ \sqlenv ->
-    f chainwebPactDb (blockEnv sqlenv) (\_ -> initSchema logger sqlenv)
+    f chainwebPactDb (blockEnv sqlenv) (\_ -> Pact5.initSchema sqlenv)
   where
     blockEnv sqlenv = BlockEnv
       (mkBlockHandlerEnv testVer testChainId (BlockHeight 0) sqlenv DoNotPersistIntraBlockWrites logger)