NoSuchTable is a new Pact error case as of Pact 5, but we never throw it. Instead, when a table is missing, sqlite throws an exception, and this is recorded as an "unknown database error" by Pact. But we can easily throw it, as shown here, and simultaneously clean up some of the more repetitious and error-prone conditionals in Pact 5's ChainwebPactDb.

Change-Id: Id000000030817f20e355d569ca29864db3d777cf