@edmundnoble 2025-01-27
Benchmarks before:
```
benchmarking PactService/Pact4/1 tx
time                 3.183 ms   (3.121 ms .. 3.254 ms)
                     0.997 R²   (0.995 R² .. 0.998 R²)
mean                 3.243 ms   (3.203 ms .. 3.276 ms)
std dev              112.7 μs   (90.03 μs .. 147.3 μs)
variance introduced by outliers: 16% (moderately inflated)

benchmarking PactService/Pact4/10 txs
time                 18.04 ms   (17.69 ms .. 18.34 ms)
                     0.998 R²   (0.997 R² .. 0.999 R²)
mean                 18.32 ms   (18.14 ms .. 18.49 ms)
std dev              423.2 μs   (347.2 μs .. 507.2 μs)

benchmarking PactService/Pact4/20 txs
time                 33.44 ms   (32.06 ms .. 34.54 ms)
                     0.996 R²   (0.989 R² .. 0.999 R²)
mean                 33.95 ms   (32.92 ms .. 34.79 ms)
std dev              1.764 ms   (1.026 ms .. 2.315 ms)
variance introduced by outliers: 13% (moderately inflated)

benchmarking PactService/Pact4/30 txs
time                 50.49 ms   (49.57 ms .. 51.47 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 49.99 ms   (49.38 ms .. 50.61 ms)
std dev              1.153 ms   (816.4 μs .. 1.759 ms)

benchmarking PactService/Pact4/40 txs
time                 64.99 ms   (63.23 ms .. 66.06 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 65.75 ms   (65.08 ms .. 66.41 ms)
std dev              1.163 ms   (769.2 μs .. 1.682 ms)

benchmarking PactService/Pact4/50 txs
time                 83.10 ms   (78.55 ms .. 88.09 ms)
                     0.996 R²   (0.989 R² .. 1.000 R²)
mean                 81.10 ms   (76.22 ms .. 83.00 ms)
std dev              4.799 ms   (1.986 ms .. 7.455 ms)
variance introduced by outliers: 10% (moderately inflated)

benchmarking PactService/Pact4/60 txs
time                 96.64 ms   (95.49 ms .. 97.43 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 96.47 ms   (95.92 ms .. 96.85 ms)
std dev              801.2 μs   (539.3 μs .. 1.169 ms)

benchmarking PactService/Pact4/70 txs
time                 111.0 ms   (108.9 ms .. 112.9 ms)
                     0.999 R²   (0.997 R² .. 1.000 R²)
mean                 113.9 ms   (112.4 ms .. 115.1 ms)
std dev              2.163 ms   (1.567 ms .. 2.922 ms)
variance introduced by outliers: 11% (moderately inflated)

benchmarking PactService/Pact4/80 txs
time                 125.0 ms   (117.4 ms .. 131.1 ms)
                     0.998 R²   (0.996 R² .. 1.000 R²)
mean                 127.7 ms   (124.9 ms .. 130.0 ms)
std dev              3.737 ms   (2.375 ms .. 5.128 ms)
variance introduced by outliers: 12% (moderately inflated)

benchmarking PactService/Pact4/90 txs
time                 147.7 ms   (145.4 ms .. 150.4 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 146.8 ms   (145.3 ms .. 147.5 ms)
std dev              1.509 ms   (848.3 μs .. 2.129 ms)
variance introduced by outliers: 12% (moderately inflated)

benchmarking PactService/Pact4/100 txs
time                 156.6 ms   (151.5 ms .. 159.7 ms)
                     0.999 R²   (0.996 R² .. 1.000 R²)
mean                 161.4 ms   (158.6 ms .. 164.1 ms)
std dev              3.921 ms   (2.487 ms .. 5.481 ms)
variance introduced by outliers: 12% (moderately inflated)

benchmarking PactService/Pact5/1 tx
time                 2.874 ms   (2.842 ms .. 2.898 ms)
                     0.999 R²   (0.998 R² .. 0.999 R²)
mean                 2.869 ms   (2.845 ms .. 2.894 ms)
std dev              83.72 μs   (70.55 μs .. 108.4 μs)
variance introduced by outliers: 14% (moderately inflated)

benchmarking PactService/Pact5/10 txs
time                 10.25 ms   (10.14 ms .. 10.43 ms)
                     0.999 R²   (0.998 R² .. 0.999 R²)
mean                 10.32 ms   (10.25 ms .. 10.39 ms)
std dev              192.2 μs   (157.8 μs .. 237.5 μs)

benchmarking PactService/Pact5/20 txs
time                 17.67 ms   (17.36 ms .. 17.96 ms)
                     0.999 R²   (0.998 R² .. 0.999 R²)
mean                 17.68 ms   (17.53 ms .. 17.81 ms)
std dev              343.7 μs   (283.0 μs .. 462.1 μs)

benchmarking PactService/Pact5/30 txs
time                 25.27 ms   (24.60 ms .. 25.76 ms)
                     0.997 R²   (0.993 R² .. 1.000 R²)
mean                 25.29 ms   (24.94 ms .. 25.53 ms)
std dev              649.6 μs   (409.0 μs .. 937.8 μs)

benchmarking PactService/Pact5/40 txs
time                 33.02 ms   (32.47 ms .. 33.62 ms)
                     0.999 R²   (0.997 R² .. 1.000 R²)
mean                 32.07 ms   (31.52 ms .. 32.53 ms)
std dev              1.236 ms   (789.7 μs .. 1.652 ms)
variance introduced by outliers: 11% (moderately inflated)

benchmarking PactService/Pact5/50 txs
time                 39.35 ms   (38.80 ms .. 40.19 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 39.13 ms   (38.38 ms .. 39.57 ms)
std dev              1.081 ms   (562.3 μs .. 1.880 ms)

benchmarking PactService/Pact5/60 txs
time                 46.71 ms   (44.47 ms .. 48.12 ms)
                     0.996 R²   (0.991 R² .. 0.999 R²)
mean                 47.64 ms   (46.74 ms .. 49.47 ms)
std dev              2.367 ms   (1.100 ms .. 3.914 ms)
variance introduced by outliers: 14% (moderately inflated)

benchmarking PactService/Pact5/70 txs
time                 54.05 ms   (52.42 ms .. 54.95 ms)
                     0.998 R²   (0.994 R² .. 1.000 R²)
mean                 52.79 ms   (50.48 ms .. 53.88 ms)
std dev              2.726 ms   (1.245 ms .. 4.363 ms)
variance introduced by outliers: 15% (moderately inflated)

benchmarking PactService/Pact5/80 txs
time                 61.38 ms   (60.33 ms .. 62.30 ms)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 60.83 ms   (60.09 ms .. 61.17 ms)
std dev              1.026 ms   (586.4 μs .. 1.618 ms)

benchmarking PactService/Pact5/90 txs
time                 69.14 ms   (68.32 ms .. 70.44 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 69.19 ms   (68.63 ms .. 69.82 ms)
std dev              975.4 μs   (734.4 μs .. 1.263 ms)

benchmarking PactService/Pact5/100 txs
time                 77.78 ms   (75.52 ms .. 80.54 ms)
                     0.997 R²   (0.992 R² .. 0.999 R²)
mean                 75.78 ms   (73.68 ms .. 77.44 ms)
std dev              3.166 ms   (1.940 ms .. 4.576 ms)
```

Benchmarks after:

```
benchmarking PactService/Pact4/1 tx
time                 3.251 ms   (3.201 ms .. 3.302 ms)
                     0.999 R²   (0.998 R² .. 0.999 R²)
mean                 3.254 ms   (3.220 ms .. 3.275 ms)
std dev              75.37 μs   (62.98 μs .. 94.99 μs)

benchmarking PactService/Pact4/10 txs
time                 18.20 ms   (17.95 ms .. 18.38 ms)
                     0.999 R²   (0.997 R² .. 1.000 R²)
mean                 18.16 ms   (17.98 ms .. 18.31 ms)
std dev              372.7 μs   (219.8 μs .. 497.3 μs)

benchmarking PactService/Pact4/20 txs
time                 33.29 ms   (32.40 ms .. 34.17 ms)
                     0.998 R²   (0.996 R² .. 1.000 R²)
mean                 34.09 ms   (33.73 ms .. 34.34 ms)
std dev              680.2 μs   (456.5 μs .. 1.069 ms)

benchmarking PactService/Pact4/30 txs
time                 50.46 ms   (49.74 ms .. 51.74 ms)
                     0.999 R²   (0.997 R² .. 1.000 R²)
mean                 49.90 ms   (48.67 ms .. 50.65 ms)
std dev              1.468 ms   (854.7 μs .. 2.171 ms)

benchmarking PactService/Pact4/40 txs
time                 66.54 ms   (65.27 ms .. 67.71 ms)
                     0.999 R²   (0.997 R² .. 1.000 R²)
mean                 64.67 ms   (63.49 ms .. 65.66 ms)
std dev              1.993 ms   (1.441 ms .. 2.934 ms)

benchmarking PactService/Pact4/50 txs
time                 80.71 ms   (77.63 ms .. 84.29 ms)
                     0.992 R²   (0.971 R² .. 1.000 R²)
mean                 82.61 ms   (80.57 ms .. 85.91 ms)
std dev              3.982 ms   (2.081 ms .. 6.514 ms)
variance introduced by outliers: 10% (moderately inflated)

benchmarking PactService/Pact4/60 txs
time                 93.08 ms   (88.99 ms .. 98.34 ms)
                     0.997 R²   (0.992 R² .. 1.000 R²)
mean                 95.84 ms   (94.29 ms .. 97.34 ms)
std dev              2.548 ms   (1.762 ms .. 3.585 ms)

benchmarking PactService/Pact4/70 txs
time                 112.0 ms   (108.6 ms .. 117.0 ms)
                     0.999 R²   (0.994 R² .. 1.000 R²)
mean                 110.2 ms   (106.7 ms .. 112.0 ms)
std dev              4.674 ms   (1.547 ms .. 6.920 ms)
variance introduced by outliers: 11% (moderately inflated)

benchmarking PactService/Pact4/80 txs
time                 125.2 ms   (120.3 ms .. 129.4 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 129.7 ms   (127.3 ms .. 131.1 ms)
std dev              2.568 ms   (1.560 ms .. 3.737 ms)
variance introduced by outliers: 12% (moderately inflated)

benchmarking PactService/Pact4/90 txs
time                 146.9 ms   (142.0 ms .. 152.4 ms)
                     0.998 R²   (0.995 R² .. 1.000 R²)
mean                 142.0 ms   (135.5 ms .. 144.8 ms)
std dev              6.073 ms   (2.199 ms .. 8.865 ms)
variance introduced by outliers: 12% (moderately inflated)

benchmarking PactService/Pact4/100 txs
time                 170.7 ms   (147.2 ms .. 188.3 ms)
                     0.995 R²   (0.987 R² .. 1.000 R²)
mean                 162.0 ms   (156.0 ms .. 166.5 ms)
std dev              7.773 ms   (5.664 ms .. 9.549 ms)
variance introduced by outliers: 12% (moderately inflated)

benchmarking PactService/Pact5/1 tx
time                 2.883 ms   (2.626 ms .. 3.185 ms)
                     0.947 R²   (0.911 R² .. 0.975 R²)
mean                 3.082 ms   (2.984 ms .. 3.205 ms)
std dev              364.3 μs   (307.4 μs .. 452.6 μs)
variance introduced by outliers: 72% (severely inflated)

benchmarking PactService/Pact5/10 txs
time                 9.955 ms   (9.681 ms .. 10.28 ms)
                     0.995 R²   (0.990 R² .. 0.998 R²)
mean                 10.18 ms   (10.03 ms .. 10.33 ms)
std dev              412.9 μs   (285.6 μs .. 659.2 μs)
variance introduced by outliers: 16% (moderately inflated)

benchmarking PactService/Pact5/20 txs
time                 17.30 ms   (16.56 ms .. 17.75 ms)
                     0.995 R²   (0.991 R² .. 0.998 R²)
mean                 17.30 ms   (17.10 ms .. 17.58 ms)
std dev              586.8 μs   (423.9 μs .. 739.9 μs)

benchmarking PactService/Pact5/30 txs
time                 24.85 ms   (24.07 ms .. 25.49 ms)
                     0.997 R²   (0.995 R² .. 0.999 R²)
mean                 24.56 ms   (24.22 ms .. 24.81 ms)
std dev              615.3 μs   (510.9 μs .. 753.7 μs)

benchmarking PactService/Pact5/40 txs
time                 31.74 ms   (30.19 ms .. 32.85 ms)
                     0.990 R²   (0.979 R² .. 0.999 R²)
mean                 31.02 ms   (30.15 ms .. 31.70 ms)
std dev              1.531 ms   (809.8 μs .. 2.120 ms)
variance introduced by outliers: 16% (moderately inflated)

benchmarking PactService/Pact5/50 txs
time                 38.88 ms   (37.83 ms .. 39.91 ms)
                     0.998 R²   (0.996 R² .. 1.000 R²)
mean                 39.10 ms   (38.75 ms .. 39.42 ms)
std dev              743.6 μs   (473.2 μs .. 1.047 ms)

benchmarking PactService/Pact5/60 txs
time                 45.72 ms   (43.83 ms .. 47.03 ms)
                     0.997 R²   (0.992 R² .. 1.000 R²)
mean                 46.31 ms   (45.34 ms .. 47.80 ms)
std dev              2.242 ms   (1.234 ms .. 3.908 ms)
variance introduced by outliers: 14% (moderately inflated)

benchmarking PactService/Pact5/70 txs
time                 51.81 ms   (49.99 ms .. 52.94 ms)
                     0.998 R²   (0.994 R² .. 1.000 R²)
mean                 52.55 ms   (51.60 ms .. 53.37 ms)
std dev              1.695 ms   (1.253 ms .. 2.401 ms)

benchmarking PactService/Pact5/80 txs
time                 59.43 ms   (58.76 ms .. 60.21 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 60.05 ms   (59.49 ms .. 61.47 ms)
std dev              1.482 ms   (673.0 μs .. 2.131 ms)

benchmarking PactService/Pact5/90 txs
time                 62.25 ms   (56.01 ms .. 70.54 ms)
                     0.979 R²   (0.948 R² .. 0.998 R²)
mean                 67.21 ms   (64.00 ms .. 68.83 ms)
std dev              3.796 ms   (2.315 ms .. 5.904 ms)
variance introduced by outliers: 17% (moderately inflated)

benchmarking PactService/Pact5/100 txs
time                 71.89 ms   (69.19 ms .. 73.18 ms)
                     0.997 R²   (0.991 R² .. 1.000 R²)
mean                 72.15 ms   (70.04 ms .. 72.85 ms)
std dev              1.876 ms   (979.8 μs .. 3.153 ms)
```